feat: add perf case tiers and raw measurement records
