+++ rstim/src/perf/cases.rs
@@ -1,0 +1,227 @@
+use serde::{Deserialize, Serialize};
+
+use crate::compiled::{
+    choose_analyzer_path, choose_sampler_path, compile_circuit, CompiledPathDecision,
+};
+use crate::ir::StimInstr;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum PerfCaseTier {
+    Gating,
+    ReportOnly,
+}
+
+impl PerfCaseTier {
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            PerfCaseTier::Gating => "gating",
+            PerfCaseTier::ReportOnly => "report_only",
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum PerfComparisonKind {
+    SamplerCompiledVsInterpreted,
+    AnalyzerCompiledVsFlattened,
+}
+
+impl PerfComparisonKind {
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            PerfComparisonKind::SamplerCompiledVsInterpreted => "sampler_compiled_vs_interpreted",
+            PerfComparisonKind::AnalyzerCompiledVsFlattened => "analyzer_compiled_vs_flattened",
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum PerfWorkload {
+    Sample,
+    Detect,
+    AnalyzeErrors,
+}
+
+impl PerfWorkload {
+    pub fn as_str(&self) -> &'static str {
+        match self {
+            PerfWorkload::Sample => "sample",
+            PerfWorkload::Detect => "detect",
+            PerfWorkload::AnalyzeErrors => "analyze_errors",
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
+pub enum PerfVariant {
+    StimCli,
+    RstimInterpreted,
+    RstimCompiled,
+    RstimAnalyzerFlattened,
+    RstimAnalyzerCompiled,
+}
+
+impl PerfVariant {
+    pub fn label(&self) -> &'static str {
+        match self {
+            PerfVariant::StimCli => "stim-cli",
+            PerfVariant::RstimInterpreted => "rstim-interpreted",
+            PerfVariant::RstimCompiled => "rstim-compiled",
+            PerfVariant::RstimAnalyzerFlattened => "rstim-analyzer-flattened",
+            PerfVariant::RstimAnalyzerCompiled => "rstim-analyzer-compiled",
+        }
+    }
+}
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum PerfCircuitSource {
+    Generator {
+        code: &'static str,
+        task: &'static str,
+        distance: usize,
+        rounds: usize,
+        noise: f64,
+    },
+    Inline {
+        text: &'static str,
+    },
+}
+
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub struct PerfBenchmarkCase {
+    pub label: &'static str,
+    pub workload: PerfWorkload,
+    pub source: PerfCircuitSource,
+    pub shots: Option<usize>,
+    pub tier: PerfCaseTier,
+    pub requires_compiled: bool,
+    pub requires_fallback: bool,
+    pub comparisons: &'static [PerfComparisonKind],
+}
+
+const SAMPLER_COMPARE: &[PerfComparisonKind] = &[PerfComparisonKind::SamplerCompiledVsInterpreted];
+const ANALYZER_COMPARE: &[PerfComparisonKind] = &[PerfComparisonKind::AnalyzerCompiledVsFlattened];
+const NO_COMPARE: &[PerfComparisonKind] = &[];
+
+pub fn benchmark_variants() -> Vec<PerfVariant> {
+    vec![
+        PerfVariant::StimCli,
+        PerfVariant::RstimInterpreted,
+        PerfVariant::RstimCompiled,
+        PerfVariant::RstimAnalyzerFlattened,
+        PerfVariant::RstimAnalyzerCompiled,
+    ]
+}
+
+pub fn benchmark_case_variants(
+    case: PerfBenchmarkCase,
+    instrs: &[StimInstr],
+) -> Result<Vec<PerfVariant>, String> {
+    let compiled = compile_circuit(instrs)?;
+    let variants = match case.workload {
+        PerfWorkload::Sample | PerfWorkload::Detect => {
+            let mut variants = vec![PerfVariant::StimCli, PerfVariant::RstimInterpreted];
+            if choose_sampler_path(&compiled) == CompiledPathDecision::FastPath {
+                variants.push(PerfVariant::RstimCompiled);
+            }
+            variants
+        }
+        PerfWorkload::AnalyzeErrors => {
+            let mut variants = vec![PerfVariant::StimCli, PerfVariant::RstimAnalyzerFlattened];
+            if choose_analyzer_path(&compiled) == CompiledPathDecision::FastPath {
+                variants.push(PerfVariant::RstimAnalyzerCompiled);
+            }
+            variants
+        }
+    };
+    Ok(variants)
+}
+
+pub fn effective_repeat_count(instrs: &[StimInstr]) -> usize {
+    effective_repeat_count_with_multiplier(instrs, 1)
+}
+
+fn effective_repeat_count_with_multiplier(instrs: &[StimInstr], multiplier: usize) -> usize {
+    let mut total = 0usize;
+    for instr in instrs {
+        if let StimInstr::Repeat { count, body } = instr {
+            let scaled = multiplier.saturating_mul(*count as usize);
+            total = total.saturating_add(scaled);
+            total = total.saturating_add(effective_repeat_count_with_multiplier(body, scaled));
+        }
+    }
+    total
+}
+
+pub fn benchmark_cases() -> Vec<PerfBenchmarkCase> {
+    vec![
+        PerfBenchmarkCase {
+            label: "rep-sample-d13-r13",
+            workload: PerfWorkload::Sample,
+            source: PerfCircuitSource::Generator {
+                code: "repetition_code",
+                task: "memory",
+                distance: 13,
+                rounds: 13,
+                noise: 0.001,
+            },
+            shots: Some(20_000),
+            tier: PerfCaseTier::Gating,
+            requires_compiled: true,
+            requires_fallback: false,
+            comparisons: SAMPLER_COMPARE,
+        },
+        PerfBenchmarkCase {
+            label: "surface-detect-d13-r13",
+            workload: PerfWorkload::Detect,
+            source: PerfCircuitSource::Generator {
+                code: "surface_code",
+                task: "rotated_memory_x",
+                distance: 13,
+                rounds: 13,
+                noise: 0.001,
+            },
+            shots: Some(10_000),
+            tier: PerfCaseTier::Gating,
+            requires_compiled: true,
+            requires_fallback: false,
+            comparisons: SAMPLER_COMPARE,
+        },
+        PerfBenchmarkCase {
+            label: "repeat-analyze-large",
+            workload: PerfWorkload::AnalyzeErrors,
+            source: PerfCircuitSource::Inline {
+                text: "REPEAT 4096 {\n    X_ERROR(0.001) 0\n    MR 0\n    DETECTOR rec[-1]\n}\n",
+            },
+            shots: None,
+            tier: PerfCaseTier::Gating,
+            requires_compiled: true,
+            requires_fallback: false,
+            comparisons: ANALYZER_COMPARE,
+        },
+        PerfBenchmarkCase {
+            label: "loss-protection-sample",
+            workload: PerfWorkload::Sample,
+            source: PerfCircuitSource::Inline {
+                text: "LOSS(1) 0\nMRL 0\nDETECTOR rec[-1]\n",
+            },
+            shots: Some(128),
+            tier: PerfCaseTier::Gating,
+            requires_compiled: false,
+            requires_fallback: true,
+            comparisons: NO_COMPARE,
+        },
+        PerfBenchmarkCase {
+            label: "repeat-analyze-stress-report",
+            workload: PerfWorkload::AnalyzeErrors,
+            source: PerfCircuitSource::Inline {
+                text: "REPEAT 8192 {\n    X_ERROR(0.001) 0\n    MR 0\n    DETECTOR rec[-1]\n}\n",
+            },
+            shots: None,
+            tier: PerfCaseTier::ReportOnly,
+            requires_compiled: true,
+            requires_fallback: false,
+            comparisons: ANALYZER_COMPARE,
+        },
+    ]
+}
